Armor Part 1
- Added most custom armor sets and their sprites

Most armor, particularly the late-game items, does not currently have a set bonus. Many set bonuses in the original mod are either yet to be implemented or are being redesigned for balancing purposes (seriously, "fatal damage causes player to revive to full health with 2x damage for a short period of time, one a minute". Who thought that was fair)